requirements.txt
textgrid
praat-parselmouth
matplotlib
seaborn
numpy
librosa
openai-whisper
